backend/requirement.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.4.0
aiofiles==23.2.1
python-multipart==0.0.9
ultralytics==8.2.103
llama-cpp-python==0.2.84
jinja2==3.1.4
orjson==3.10.6
websockets==12.0
tqdm==4.66.4
pytest==8.2.1
httpx==0.27.0
gpt4all==2.7.0
